requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
SQLAlchemy==2.0.36
alembic==1.13.3
psycopg[binary]==3.2.10    # <-- usar a série 3.2.x
python-dotenv==1.0.1
